refactor(client/go)!: delete legacy configyaml and configregistry parsers
The composer-based system (composer + netconfig + connmeta) replaces
them entirely.

Also restores errorStreamDialer (a test-only fake dialer) inside
iptable_config_test.go: it previously lived in registry_test.go and
was still needed by iptable_config_test.go's dial-dispatch tests.
