+++ client/go/outline/configregistry/outline_dns.go
@@ -30,7 +30,7 @@ import (
 )
 
 // A list of public DNS resolvers that the VPN can use.
-var outlineDNSResolversV2 = []netip.AddrPort{
+var outlineDNSResolvers = []netip.AddrPort{
 	netip.MustParseAddrPort("1.1.1.1:53"),        // Cloudflare
 	netip.MustParseAddrPort("9.9.9.9:53"),        // Quad9
 	netip.MustParseAddrPort("208.67.222.222:53"), // OpenDNS
@@ -40,7 +40,7 @@ var outlineDNSResolversV2 = []netip.AddrPort{
 // A hard-coded link-local address for DNS interception.
 //
 // TODO: make this configurable via a new VpnConfig
-var linkLocalDNSV2 = netip.MustParseAddrPort("169.254.113.53:53")
+var linkLocalDNS = netip.MustParseAddrPort("169.254.113.53:53")
 
 // NewOutlineDNSTransport applies Outline's DNS policy to a built
 // transport: it intercepts DNS at a link-local address over TCP and
@@ -50,11 +50,11 @@ var linkLocalDNSV2 = netip.MustParseAddrPort("169.254.113.53:53")
 // packet relay, and the network-change callback.
 func NewOutlineDNSTransport(sd transport.StreamDialer, pl transport.PacketListener) (transport.StreamDialer, packetrelay.PacketRelay, func(), error) {
 	// Randomly selects a DNS resolver for the VPN session
-	remoteDNS := outlineDNSResolversV2[rand.IntN(len(outlineDNSResolversV2))]
+	remoteDNS := outlineDNSResolvers[rand.IntN(len(outlineDNSResolvers))]
 
-	// Intercept DNS for StreamDialer: remap TCP connections to linkLocalDNSV2 → remoteDNS.
+	// Intercept DNS for StreamDialer: remap TCP connections to linkLocalDNS → remoteDNS.
 	sdForward := func(ctx context.Context, addr string) (transport.StreamConn, error) {
-		if dst, err := netip.ParseAddrPort(addr); err == nil && dst.Addr().Unmap() == linkLocalDNSV2.Addr() && dst.Port() == linkLocalDNSV2.Port() {
+		if dst, err := netip.ParseAddrPort(addr); err == nil && dst.Addr().Unmap() == linkLocalDNS.Addr() && dst.Port() == linkLocalDNS.Port() {
 			addr = remoteDNS.String()
 		}
 		return sd.DialStream(ctx, addr)
@@ -70,14 +70,14 @@ func NewOutlineDNSTransport(sd transport.StreamDialer, pl transport.PacketListen
 	if err != nil {
 		return nil, nil, nil, fmt.Errorf("failed to create DNS PacketRelay: %w", err)
 	}
-	relayForward := dnsintercept.NewInterceptDNSPacketRelay(dnsListener, baseListener, linkLocalDNSV2, remoteDNS)
+	relayForward := dnsintercept.NewInterceptDNSPacketRelay(dnsListener, baseListener, linkLocalDNS, remoteDNS)
 	// Truncate relay: intercept DNS at link-local address, return truncated response (forces TCP retry).
 	// Non-DNS traffic passes through to baseListener.
 	dnsTruncRelay, err := dnstruncate.NewPacketRelay()
 	if err != nil {
 		return nil, nil, nil, fmt.Errorf("failed to create DNS truncate relay: %w", err)
 	}
-	relayTrunc := dnsintercept.NewInterceptDNSPacketRelay(dnsTruncRelay, baseListener, linkLocalDNSV2, remoteDNS)
+	relayTrunc := dnsintercept.NewInterceptDNSPacketRelay(dnsTruncRelay, baseListener, linkLocalDNS, remoteDNS)
 	// Delegate relay starts with truncate (UDP unverified), switches to forward when UDP is healthy.
 	relayMain, err := packetrelay.NewDelegatePacketRelay(relayTrunc)
 	if err != nil {
